Update CMD_INFO.cs (#60)
Discovered that these functions with two GFX entries use the first entry when character is facing right and the second entry when character is facing left. This could very well apply to all functions with two GFX entries, but I only know for sure it works that way for these two.